+++ module/Core/src/ShortUrl/Model/Validation/CustomSlugValidator.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Shlinkio\Shlink\Core\ShortUrl\Model\Validation;
 
 use Laminas\Validator\AbstractValidator;
-use Shlinkio\Shlink\Core\Options\UrlShortenerOptions;
+use Shlinkio\Shlink\Core\Config\Options\UrlShortenerOptions;
 
 use function is_string;
 use function strpbrk;