+++ module/Core/test/Matomo/MatomoVisitSenderTest.php
@@ -11,6 +11,7 @@ use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Shlinkio\Shlink\Common\Util\DateRange;
+use Shlinkio\Shlink\Core\Config\Options\UrlShortenerOptions;
 use Shlinkio\Shlink\Core\Matomo\MatomoTrackerBuilderInterface;
 use Shlinkio\Shlink\Core\Matomo\MatomoVisitSender;
 use Shlinkio\Shlink\Core\ShortUrl\Entity\ShortUrl;
@@ -36,7 +37,7 @@ class MatomoVisitSenderTest extends TestCase
 
         $this->visitSender = new MatomoVisitSender(
             $this->trackerBuilder,
-            new ShortUrlStringifier(['hostname' => 's2.test']),
+            new ShortUrlStringifier(new UrlShortenerOptions(defaultDomain: 's2.test')),
             $this->visitIterationRepository,
         );
     }